football-bot/requirements.txt
telethon
feedparser
python-dotenv